chore: 600 rpm for embedding by default
