+++ crates/http-api-bindings/src/embedding/mod.rs
@@ -1,20 +1,15 @@
 mod llama;
 mod openai;
+mod rate_limit;
 mod voyage;
 
-use core::{panic, time};
-use std::{
-    sync::Arc,
-    task::{Context, Poll},
-};
+use core::panic;
+use std::sync::Arc;
 
-use async_trait::async_trait;
-use futures::future::BoxFuture;
 use llama::LlamaCppEngine;
+use rate_limit::RateLimitedEmbedding;
 use tabby_common::config::HttpModelConfig;
 use tabby_inference::Embedding;
-use tokio::sync::Mutex;
-use tower::{Service, ServiceBuilder, ServiceExt};
 use tracing::debug;
 
 use self::{openai::OpenAIEmbeddingEngine, voyage::VoyageEmbeddingEngine};
@@ -23,10 +18,14 @@ pub async fn create(config: &HttpModelConfig) -> Arc<dyn Embedding> {
     let rpm = if let Some(limit) = &config.request_limit {
         limit.request_per_minute
     } else {
-        0
+        debug!(
+            "No request limit specified for model {}, defaulting to 600 rpm",
+            config.kind
+        );
+        600
     };
 
-    let embedding = match config.kind.as_str() {
+    let embedding: Arc<dyn Embedding> = match config.kind.as_str() {
         "llama.cpp/embedding" => {
             let engine = LlamaCppEngine::create(
                 config
@@ -37,7 +36,6 @@ pub async fn create(config: &HttpModelConfig) -> Arc<dyn Embedding> {
             );
             Arc::new(engine)
         }
-        "ollama/embedding" => ollama_api_bindings::create_embedding(config).await,
         "openai/embedding" => {
             let engine = OpenAIEmbeddingEngine::create(
                 config
@@ -49,6 +47,7 @@ pub async fn create(config: &HttpModelConfig) -> Arc<dyn Embedding> {
             );
             Arc::new(engine)
         }
+        "ollama/embedding" => ollama_api_bindings::create_embedding(config).await,
         "voyage/embedding" => {
             let engine = VoyageEmbeddingEngine::create(
                 config.api_endpoint.as_deref(),
@@ -69,68 +68,7 @@ pub async fn create(config: &HttpModelConfig) -> Arc<dyn Embedding> {
         ),
     };
 
-    if rpm > 0 {
-        debug!(
-            "Creating rate limited embedding with {} requests per minute",
-            rpm,
-        );
-        Arc::new(
-            RateLimitedEmbedding::new(embedding, rpm)
-                .expect("Failed to create rate limited embedding"),
-        )
-    } else {
-        embedding
-    }
-}
-
-struct EmbeddingService {
-    embedding: Arc<dyn Embedding>,
-}
-
-impl Service<String> for EmbeddingService {
-    type Response = Vec<f32>;
-    type Error = anyhow::Error;
-    type Future = BoxFuture<'static, Result<Self::Response, Self::Error>>;
-
-    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn call(&mut self, prompt: String) -> Self::Future {
-        let embedding = self.embedding.clone();
-        Box::pin(async move { embedding.embed(&prompt).await })
-    }
-}
-
-pub struct RateLimitedEmbedding {
-    embedding: Arc<Mutex<tower::util::BoxService<String, Vec<f32>, anyhow::Error>>>,
-}
-
-impl RateLimitedEmbedding {
-    pub fn new(embedding: Arc<dyn Embedding>, rpm: u64) -> anyhow::Result<Self> {
-        if rpm == 0 {
-            anyhow::bail!(
-                "Can not create rate limited embedding client with 0 requests per minute"
-            );
-        }
-
-        let service = ServiceBuilder::new()
-            .rate_limit(rpm, time::Duration::from_secs(60))
-            .service(EmbeddingService { embedding })
-            .boxed();
-
-        Ok(Self {
-            embedding: Arc::new(Mutex::new(service)),
-        })
-    }
-}
-
-#[async_trait]
-impl Embedding for RateLimitedEmbedding {
-    async fn embed(&self, prompt: &str) -> anyhow::Result<Vec<f32>> {
-        let mut service = self.embedding.lock().await;
-        let prompt_owned = prompt.to_string();
-        let response = service.ready().await?.call(prompt_owned).await?;
-        Ok(response)
-    }
+    Arc::new(
+        RateLimitedEmbedding::new(embedding, rpm).expect("Failed to create rate limited embedding"),
+    )
 }
