+++ crates/http-api-bindings/src/embedding/openai.rs
@@ -12,7 +12,7 @@ pub struct OpenAIEmbeddingEngine {
 }
 
 impl OpenAIEmbeddingEngine {
-    pub fn create(api_endpoint: &str, model_name: &str, api_key: Option<&str>) -> impl Embedding {
+    pub fn create(api_endpoint: &str, model_name: &str, api_key: Option<&str>) -> Self {
         let config = OpenAIConfig::default()
             .with_api_base(api_endpoint)
             .with_api_key(api_key.unwrap_or_default());
