+++ crates/http-api-bindings/src/lib.rs
@@ -5,13 +5,6 @@ mod embedding;
 pub use chat::create as create_chat;
 pub use completion::{build_completion_prompt, create};
 pub use embedding::create as create_embedding;
-use reqwest::Response;
-use serde_json::Value;
-
-#[async_trait::async_trait]
-pub trait HttpClient: Send + Sync {
-    async fn post(&self, url: &str, body: Value) -> anyhow::Result<Response>;
-}
 
 fn create_reqwest_client(api_endpoint: &str) -> reqwest::Client {
     let builder = reqwest::Client::builder();
